refactor(cli): context providers use dedicated, provided messaging class (#277)
Introduces a dedicated messaging interface for Context Providers. This
is a reduced interface to allow Context Providers send debug messages
and in rare cases info messages.

Use of `ioHelper` in the `provideContextValues` function is currently
still optional until the CloudExecutable supports it as well.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>