chore(tmp-toolkit-helpers): `formatStackDiff` and `formatSecurityDiff` moved to tmp-toolkit-helpers (#278)
moves two functions that are relevant to diff into
`tmp-toolkit-helpers`. this will facilitate the adoption of `diff` in
the programmatic toolkit as it will now be able to reuse these
functions.

existing tests pass to ensure no breakage. additional tests have been
added to `tmp-toolkit-helpers` to test `formatXxxDiff`, which were
previously not directly tested.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>