refactor(tmp-toolkit-helpers): DiffFormatter class (#283)
the point of this PR is to refactor `formatStackDiff` and
`formatSecurityDiff` into the same class. there's more to do to simplify
these functions, but i'm making small refactors a priority.

the achievement of this PR is to get rid of the giant laundry list of
properties that go into each function in favor of a property bag, and to
store the repeated properties within a class.

this is tested by current unit tests -- as long as they still pass, this
refactor should have not changed any functionality.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>