feat(cli): CcApi context provider can be configured to fail if listing does not find a specific count of resources (#251)
Fixes #257

As described in the issue, `CcApiContextProviderPlugin` should have the
way to ensure the result has exact one resource.

This PR adds the `expectedMatchCount` option to restrict the length of
results in `listResources()`.

Unit tests are added and modified ensure the changes.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Rico Huijbers <rix0rrr@gmail.com>