feat(toolkit-lib): default to using a non interactive IoHost (#306)
Replaces the current default `IoHost` which is the same as the one used
by the CLI with a strictly non-interactive version. This is more in
accordance with the goals of a library providing actions.

Fixes #157
Closes #289 

BREAKING CHANGE: This change updates the default `IoHost` implementation
used by `Toolkit` to a version that is strictly non-interactive, i.e.
there is no expectation anymore that users will respond to command-line
prompts. To restore previous behavior, you will can provide a custom
`IoHost` implementation to your `Toolkit` instance. You may consider
extending the new `NonInteractiveIoHost` class with desired interactive
prompts.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license