feat(toolkit): watch operation can be stopped (#308)
The `toolkit.watch()` operation used to start a filesystem watcher in
the background and immediately return. The caller would continue running
to its end, while the filesystem watcher would keep the node process
alive.

Instead, now return an `IWatcher` object that can be disposed to stop
the watching, and can be used to wait for the watcher to stop.

Closes https://github.com/aws/aws-cdk-cli/issues/299

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license