refactor(cli): inject CLI version as a parameter to Notices (#323)
`Notices` will eventually serve both toolkit-lib and cli. Currently the
class has a hard-coded "cli version" - but really the value is reading
the version from whatever the current package is. Once we start using
this file in another package, it will be wrong. For now, let's just
remove the hard dependency on cli version inside the class. In future it
will be extended.

- Remove direct dependency on versionNumber import in Notices class
- Add cliVersion as a required parameter in NoticesProps interface
- Store cliVersion as a class property and use it in display method
- Improve testability by removing the need to mock version functions

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license