refactor(toolkit-lib): diff action uses helper methods (#301)
refactors the diff action to use helper methods and in general cleans up
the code a lot. introduces an internal concept of `TemplateInfo` that
standardizes the to-be-diffed information, so there's not a bunch of
repetitive code

existing tests continue to succeed. 

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>
Co-authored-by: Eli Polonsky <Eli.polonsky@gmail.com>