chore: update cdklabs-projen-project-types to 0.2.14 and set maxWorkers to 150 (#330)
This PR updates the cdklabs-projen-project-types package to version
0.2.14 and sets the maxWorkers option on CdkCliIntegTestsWorkflow to 150
as described in
https://github.com/cdklabs/cdklabs-projen-project-types/pull/841.

This change sets an appropriate maxWorkers value (150) that's optimized
for the GitHub workflow runner environment. This ensures we're making
efficient use of the available resources during integration tests.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license