refactor(toolkit): move aws-auth and plugin to shared package (#327)
This intentionally only moves the source, and not the tests. Tests will
be moved later to keep the diff size managable. Code is still tested as
the tests are still running from `aws-cdk` package.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license