refactor(toolkit): pull out interesting parts of cxapp (#329)
There is some code in the cxapp api that is needed in both CLI and
toolkit-lib. This PR breaks the shared code out and moves it into the
temp helper package. The remaining part of cxapp is "downgraded" from
being a reusable api to being CLI only.

Also includes some more hard copies of legacy exports, to make future
life easier.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license