refactor(toolkit): move context providers (#337)
Moves the context providers to the temporary helper package. Tests will
be moved later to keep the diff size manageable. Code is still tested as
the tests are still running from `aws-cdk` package.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license