refactor(toolkit): move the rest (#338)
Moves the the rest of shared toolkit code to the helper package and
removes the `toolkit-lib`s dependency on the CLI. Tests will be moved
later to keep the diff size manageable. Code is still tested as the
tests are still running from aws-cdk package.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>