chore: upgrade to alpha `jest-environment-node` (#343)
We need a prerelease version of a Jest package in order to have
`Symbol.asyncDispose` available in Jest worker processes. Apparently
these global symbols have to be copied over explicitly when Jest is
preparing a test environment.

See https://github.com/jestjs/jest/issues/14874

This does not affect "production" consumers as we have a polyfill
package that we load when the library is imported, but this polyfill
package is not loaded during test execution, which is why we need the
jest update.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license