chore: don't search for coverage files (#346)
The `codecov/codecov-action` will accept a set of coverage files to
upload, and in addition will search for coverage files to upload as
well.

From the logs you can see that it also finds coverage files in the `.nx`
cache directory.

Disable the search feature so that it only accepts the list of files we
are already passing into it.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license