feat(refactor): resources with the same physical ID are considered equivalent (#360)
Incorporate the physical ID into the computation of resource digests.
This gives the user more flexibility: if a resource with physical ID set
is moved and updated at the same time, the algorithm can still detect
that it is the same resource.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license