feat(toolkit-lib): fallBackToTemplate option for diffs (#354)
The primary motivation for this PR is
https://github.com/aws/aws-cdk/issues/28753 and the ensuing PR
https://github.com/aws/aws-cdk/pull/32830.

When implementing diff in the programmatic toolkit library we included
the diff method options, and this PR adds the `fallBackToTemplate`
property to the `ChangeSet` mode. The idea is that people can specify
`fallBackToTemplate=false` to explicitly fail if we cannot successfully
create the changeset. The current behavior is to fallback to the
template-only method.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license