chore(toolkit-lib): diff test fails if AWS credentials set (#381)
The `diff.test` times out trying to get the current account from actual
AWS credentials, if they happen to be in the environment where unit
tests are run.

Make sure a fake SDK provider is used that doesn't do any of that stuff.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license