feat(toolkit-lib): make the plugin host configurable (#375)
This makes the plugin host configurable for toolkit users:

## Design decisions

- Every `Toolkit` instance gets a fresh `PluginHost` by default (not a
global singleton `PluginHost`)
- ...in fact, the concept of a global singleton `PluginHost` gets fully
moved to the CLI. That concept disappears from the toolkit and the
helpers, and all functions that used to implicitly assume a global
plugin host now get an explicit parameter, that defaults to a fresh
plugin host if not supplied.

These choices lead to a lot of downstream test changes.

- Plugin module resolution, module deduplication and logging was moved
from the CLI into the plugin host.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license