feat(toolkit-lib): make base credentials configurable (#388)
Previously, the Toolkit would always use AWS CLI-compatible base
credentials. This is now configurable:

```ts
const toolkit = new Toolkit({
  sdkConfig: {
    baseCredentials: BaseCredentials.custom(...),
  },
});
```

## Design

The `BaseCredentials` (abstract) class is responsible for producing the
following 2 bits of information:

- A credential provider
- A default region

These will be used to initialize an `SdkProvider`, which will then
proceed to use that information:

- To inform the CDK app about the desired target environment
- After synthesis and during lookup:
  - use those credentials directly; or
  - use those credentials to assume roles; or
  - use available plugins to obtain credentials

`requestHandler` (the proxy agent plus some SDK settings) used to be
produced by the `SdkProvider` itself, but it is now produced by the
`Toolkit` and part of the "services" that get passed in. That way, it is
both available to the `AwsCliCompatible` base credentials to initialize
the STS client, as well as to the `SdkProvider` class that now also gets
instantiated by the `Toolkit`.


![image](https://github.com/user-attachments/assets/4ee1656b-d54f-4d1d-8ac7-77cd75c145b2)


## Supporting changes

- Many of the supporting positional arguments to `SdkProvider` have been
grouped into `SdkProviderServices`.
- Change how we mock the SDK in a number of tests (instead of mocking
the SDK Provider, mock the SDK)

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license