fix(cli): remove unnecessary `dom` lib from default TypeScript config created with `cdk init` (#295)
Fixes #294

This change also removes `"dom"` from the `tsconfig.json` files of
additional packages that are all intended to run in a Node.js
environment, not in the browser.

Since these packages are not expected to use any web runtime-specific
APIs, removing `"dom"` helps prevent accidental use of browser-only
types across the codebase.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>