chore: move over integ tests (#389)
Keeping the CLI integ tests in a [separate
repository](https://github.com/aws/aws-cdk-cli-testing) was a
theoretically good idea and a practically annoying idea.

Let's move them in here, this is much easier to work with.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>