chore(cli-integ): node-gyp fails on newer python versions (#407)
The `cli-integ` tests depend on the `node-pty` package for testing of
interactive terminals. This package needs to be build at install time
using `node-gyp`. Because previously `cli-integ` tests dependet on an
old version of `npm`, we got an old version of `node-gyp` which is
incomaptible with recent python versions.

To fix this, we upgrade the dependency to a more recent version of
`npm`.

Fixes failing Python package release.

Proof this fix works:
https://github.com/aws/aws-cdk-cli/actions/runs/14619849549/job/41016752989?pr=406

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license