refactor(toolkit-lib): remove `requireApproval` option from diff (#372)
`requireApproval` is deprecated and has been moved to the `CliIoHost`.
this PR removes the last place where `requireApproval` was necessary in
`tmp-toolkit-helpers`, in `formatSecurityDiff`. `formatSecurityDiff` now
makes no _decisions_ on what to print; it returns the diff and a
`permissionChangeType` for the `IoHost` to interpret. this is a slight
behavior change, so all consumers of `formatSecurityDiff have been
updated accordingly (`deploy` and `diff` in `toolkit-lib` and in the
CLI).

`requireApproval` is now a vestigial structure only in use in the CLI. 

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>