chore: log integ test output to the file system (#416)
Write the (huge) text logs of integ test files to the file system if
configured to do so (with `$INTEG_LOGS`), and use that feature in the
GitHub logs to publish the logs as an artifact.

At the same time, also start playing with Markdown reports of GitHub
actions by writing to `$GITHUB_STEP_SUMMARY` to see what that looks
like.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>