feat(refactor): support for stack filtering (#410)
Users can now pass the stacks they want to be considered for
refactoring:

    cdk refactor StackA StackB ...

The result will include any mapping that contains at least one of the
selected stacks. If no pattern is passed, all stacks are considered.

In this PR I also included the `refactor` command in the command line
parsing list, which was missing. It was already working without it, but
the command wouldn't show up in the help page, and it wouldn't recognize
positional arguments. But the command itself (and its options) could be
already used.

Closes https://github.com/aws/aws-cdk-cli/issues/363.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Kaizen Conroy <36202692+kaizencc@users.noreply.github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>