chore: make logs artifacts unique per matrix job (#426)
We used to try to upload all logs under the same artifact name, but we
need to make them explicitly unique per matrix job otherwise the names
collide.

Also include a link to the artifact name in the summary, just for kicks.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>