chore: allow running integ tests on all branches (#431)
We used to only run integ tests for PRs targeting `main`.

That makes it impossible to run integ tests for stacked PRs, which is
inconvenient. Remove the restriction and run integ tests on all PRs.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>