refactor(toolkit-lib): remove unnecessary bundling (#435)
We previously bundled some entrypoints and their type declarations
because they containted code that lived in the
not published shared helper package. With that package now remove,
bundling is no longer needed.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license