fix(integ-tests): tests need to `mv`ed before they can be run (#439)
For historical reasons of working with Jest 27 or lower, we needed to
`mv` the `integ-tests` out of an install directory where they would be
left by an `npm install`. The reason was that Jest 27 or lower would
always ignore all `node_modules` direcotries, and so could not be
configured to run tests from an install location.

Since we are on Jest 29, we can now lift this limitation with the right
configuration options. We can run the tests directly from the
`node_modules` directory and don't have to fiddle around with moving
files anymore.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>