feat(refactor): ability to exclude selected resources from refactoring (#421)
Allow the user to select resources to be skipped for refactoring. There
are two ways to do this:

- By adding the `aws:cdk:skip-refactor = true` metadata to each resource
to be skipped, in the cloud assembly manifest.
- By passing a skip fie via command line or the toolkit library API. The
file should be in JSON format and contain a list of destination
locations. Entries in this list can be either in the form `"<stack
name>.<logical ID>"` or a construct path.

The main abstraction is the `SkipList` interface, used by the mapping
detection algorithm to query for mappings that should be excluded.

Closes https://github.com/aws/aws-cdk-cli/issues/412,
https://github.com/aws/aws-cdk-cli/issues/379

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license