fix(cli): `cdk notices` ignores failures and calling cli with `--no-notices` makes an unexpected network call (#399)
Fixes #370 
Fixes #440

Previously when the `--no-notices` option was passed, the CLI would
still attempt to refresh notices in the background. This is unnecessary
and an unexpected behavior to users, especially when running the CLI in
environments with no internet access.

On the other hand, running the explicit `cdk notices` command did
previously not fail if the command failed to refresh notices. Instead
the command would pretend no notices are found and complete
successfully.

This PR updates the code to not refresh notices when notices are
disabled. Unless the explicit `cdk notices` command is called, which
will always refresh and now error if fetching notices fails. It also
removes a double refresh of the notices (which didn't have an effect
because they were cached, but still).

`cdk notices` with an error now looks like this:

<img width="775" alt="image"
src="https://github.com/user-attachments/assets/5c09263c-a775-49c9-8dcf-e4feeefb10ac"
/>

Before the command would incorrectly and silently succeed and not show
any notices:

<img width="118" alt="image"
src="https://github.com/user-attachments/assets/dd93c1bf-4b6a-4fa9-8c9a-1fbacccc0156"
/>

Only when run with the `-v` flag, it would hint at the failure:

```console
$ cdk notices -v
...
[12:38:29] Could not refresh notices: Error: getaddrinfo ENOTFOUND cli.cdk.dev-tools
```
---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license