chore(integ-tests): decouple CLI and Library sources, improve documentation (#425)
We used to make assumptions in the integ tests that the CLI and library
versions were the same unless specified otherwise, and that they both
would come from the same repo.

These assumptions are no longer true: the CLI and the library have their
own version numbers, and only the CLI can come from a source repo (or at
least, that's how it works in practice).

Tear apart `ReleasePackageSource` and `RepoPackageSource` into
(currently):

- `CliNpmSource` / `CliRepoSource`
- `LibraryNpmSource`

Make a more clear naming distinction between the part that runs in the
runner and the part that runs in the test.

Explain in documentation what the options are, and how the use of the
given component is achieved at runtime.

Make the argument names and default values a little more sensible given
the new source locations.

CHANGES:

- Add `--cli-version` to mirror `--framework-version`.
- Default to `--auto-source` if no CLI version given.
- Add `-F` alias to specify a single test file.
- Resolve the framework version immediately.
- Deprecate `--use-source` and replace with `--cli-source`.
- Deprecate `--auto-source` and replace with `--cli-source=auto`.
- Change to a different `yargs` version and twiddle all the options
slightly differently to enable strict parsing mode.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

Signed-off-by: github-actions <github-actions@github.com>