chore: also write logs if integ tests fail (#456)
Currently, if the integ test suite fails, the step summary doesn't get
written and the logs don't get uploaded.

Add

```
if: always()
```

To the steps to make sure they always run, even if the job is failing.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license