feat(integ-tests): add integ tests for the toolkit library (#430)
Add integration tests for the toolkit library.

Introduces the toolkit library as a new "component under test". 

Because we want to promote writing integ tests against the toolkit
library the same way we would write normal production code, the setup is
like this:

- The toolkit lib is installed as a `devDependency` into the integ test
package. This will allow writing tests as if the toolkit lib is "just" a
dependency.
- At runtime, a version of the toolkit library can be selected and it
will be installed into the dependency closure by the runner before the
tests are run, so that the imports will still function properly.

To signal this, we'll mark `@aws-cdk/toolkit-lib` as an
`optionalDependency` as well.

(PR includes https://github.com/aws/aws-cdk-cli/pull/444)

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

Signed-off-by: github-actions <github-actions@github.com>