chore: reorg eslint rules into separate files (#466)
The rules file got a bit large and hard to maintain/understand what's
going on. In this change we are splitting the file up into sensible
categories.

Also includes a projen upgrade to uses `@stylistic` for formatting rules
and fixes a couple of rules to not use an array when they prefer to not
use one (based on VSCode eslint plugin warnings). Places where some of
the affected rules are disabled needed updating as well.

There are a bunch of small formatting changes to type statements. These
didn't get picked up previously because the built-in eslint rules don't
support TypeScript, but the `@stylistic` ones do.

The lack of any other changes to the codebase proves that these fixes
are alright.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>