feat(cli): configurable timeout for ECS hotswap operation (#417)
During hotswap deployment, we wait for the ECS service to become stable:


https://github.com/aws/aws-cdk-cli/blob/f6bd7b9056186c00160ece09ac2d4ddc2ad72f35/packages/%40aws-cdk/tmp-toolkit-helpers/src/api/hotswap/ecs-services.ts#L153-L156

The timeout is hardcoded to 600 seconds, which is a very long time to
**have** to wait. This PR add an additional property to the already
existing ECS overrides bag, to configure this timeout.

### Notes

Our integ tests currently suffer from this because they needlessly wait
the hard-coded 10 minute timeout in a test that asserts ECS updates
error handling. With this option, we set the operation timeout to 10
seconds and reduce the total integ suite duration **from 17min to
9min**.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Rico Hermans <rix0rrr@gmail.com>