fix(cloudformation-diff): `@aws-sdk/client-cloudformation` should be a peer dependency (#478)
The function signature of `fullDiff` depends on the
`DescribeChangeSetOutput` type from `@aws-sdk/client-cloudformation`.
Therefore this package should be a peer dependency, not a dev
dependency.

Updates the dependency type, and adds the minimal version as a devDep to
ensure we stay compatible. I needed to add `0BSD` into the license
checker since one of the transitive dependencies use it. It is an
approved license.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license