refactor(toolkit-lib): remove duplicate SdkLogger and use copy of interface (#467)
We had two version of an `SdkLogger` that logs to an IoHost. This was
because historically they lived in different packages. Align both
versions to use the CLI one - because that is definitely safe and battle
tested.

Additionally we move away from using the `Logger` interface provided by
`@smithy/types`. This is in preparation of fully removing this
dependency.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license