feat: generate stack definitions for refactoring (#477)
Having a list of deployed stacks, and the mappings (either computed or
user provided), we have to generate a list of stack definitions to be
sent to the `createStackRefactor` API.

The function `generateStackDefinitions` is not yet used anywhere, other
than unit tests. Future PRs will use it to make the API call.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license