chore(toolkit-lib): use `isolatedModules` from tsconfig instead of `ts-jest` (#482)
`isolatedModules` is a tsconfig that can help improve the performance of
tools interpreting .ts files in single-file transpilation processes. In
our case that's what `ts-jest` is doing. We are currently getting
something like this behavior by settings ts-jest's `isolatedModules`
setting (and yes: that's a different setting with the same goal).
However we are also getting a bunch of warnings about this setting from
`ts-jest` as it will be deprecated in the next version.

This change serves as a testing ground to see if and how we can enable
tsc's `isolatedModules` for all packages. It enables TypeScript's
`isolatedModules` and disables `ts-jest`'s `isolatedModules`. To make
this change work, a few minor changes had to be made to the
`toolkit-lib` code. I then manually benchmarked the time it takes for
tests to run, and there seems to be no performance difference.

Reference: https://www.typescriptlang.org/tsconfig/#isolatedModules

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license