chore: require type imports in published packages to be from declared dependencies (#472)
When we import a type, this import may become part of the public
declarations of a package. Previously we allowed types to be imported
from devDeps. This can cause consuming TypeScript packages to fail
compilation if they have `skipLibCheck: false` set (which is the
default).

To resolve this, we are now enforcing all type imports to be from a
declared dependency. The recommendation is to declare a very wide peer
dependency, if the package is a types only package.

Note that not all imports of types will actually end up in the published
declaration files. This only happens if the imported type is part of an
exported API for that file, e.g. part of a function signature.
Unfortunately there is no good way of checking just this case with
eslint, so we do the next best thing.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license