feat(cli): use explicit mapping when provided by the user (#460)
This PR introduces two new options to the `refactor` command:
`--mapping-file` and `--revert`. When a mapping file is provided, the
CLI won't do try to compute the refactor mappings, and just use what the
user instructed it to. The `--revert` can only be used in conjunction
with `--mapping-file`, and modifies its behavior by swapping source and
destination.

Closes #378.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: Rico Hermans <rix0rrr@gmail.com>
Co-authored-by: github-actions <github-actions@github.com>