chore(toolkit-lib): add caret to aws-cdk/cli-plugin-contract devDep to unblock release (#496)
At the end of the release process build we are restoring versions for
local monorepo packages. This is currently hard-coded to always include
a caret `^`. While not ideal, it doesn't matter in our case. Before we
do a proper fix, unblock the release by just adding the caret.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license