fix(cli): gc does not delete isolated assets when rollback-buffer-days is set (#502)
The `Date` class allows the following for values: `new Date(value:
number | string | Date)`, but sometimes does not work as expected.
That's essentially what we are doing because we are storing the string
value of the date in the tags and the conversion back to a date doesn't
work correctly.

A deep dive into the issue with Dates:

```ts
console.log(Date.now()) // 1747346265800
console.log(Date(1747346265800)) // Thu May 15 2025 18:06:11 GMT-0400 (Eastern Daylight Time)
console.log(Date('1747346265800')) // Thu May 15 2025 18:05:57 GMT-0400 (Eastern Daylight Time)
console.log(new Date(1747346265800)) // 2025-05-15T21:57:45.800Z
console.log(new Date('1747346265800')) // Invalid Date
```

I seem to have been unlucky with how I tried to create the date from a
string. Also, the resulting dates with a number of milliseconds versus a
string of milliseconds is _slightly_ different, 18:06:11 versus
18:05:57...

This _was_ attempted to be unit tested but in the unit test I mistakenly
mocked the tag to be an ISO String which _is_ a string so it had no
problem getting converted into a Date. This test has been updated.

Because we pessimistically handle errors; we just treated this as a
nondeletable asset. I have added an integ test that tests this scenario
and confirmed on my own local set up that this fixes the issue.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license