fix(cdk-assets): indicate correct docker command that failed to execute (#504)
Users may override the executable used (instead of `docker`) by setting
the CDK_DOCKER environment variable, but the message of the error thrown
when such an executable does not exist was not similarly dynamic. Align
the contents of the message with the executable invoked.

Fixes #282.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: Ian Kerins <git@isk.haus>
Co-authored-by: Momo Kornher <mail@moritzkornher.de>