fix(toolkit-lib): dependency constraints are too strict (#514)
Due to a bug, we set constraints for dependency for other monorepo
packages to "any future minor version" (i.e. `^1.2.3`) forcing consumers
to always use latest packages with the toolkit-lib.

What we wanted to do is constrain the use to "any minor version",
including already published ones, i.e. `^1`.

This change needed an upgrade of our projen Monorepo project type, which
is also part of this PR. The API changed slightly here, hence the update
of other values.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license